requirements.txt
crewai
requests
litellm
python-dotenv
numpy
psycopg2-binary
SQLAlchemy
streamlit